modularised guardian api exported all functions for testing relates to #1
